Resolve promise when finished in both directions
We are now resolving the promise only on destroy.
The promise would never be rejected.

The reason for this change is a undesired behavior when using
the infinite scroll in both directions.

Since only one promise is returned for both directions,
when items are all loaded in one of the directions and we resolve
the promise, progress notifications are no longer fired for the
other direction.